Correct indentation for multi namespace.
Ref. #1813
