+++ CMakeLists.txt
@@ -263,6 +263,7 @@ set(uncrustify_sources
   src/compat_win32.cpp
   src/detect.cpp
   src/enum_cleanup.cpp
+  src/flag_parens.cpp
   src/indent.cpp
   src/keywords.cpp
   src/lang_pawn.cpp
@@ -323,6 +324,7 @@ set(uncrustify_headers
   src/chunk_list.h
   src/enum_cleanup.h
   src/enum_flags.h
+  src/flag_parens.h
   src/language_tools.h
   src/ListManager.h
   src/logger.h
