+++ tests/cli/output/universalindent.cfg
@@ -2209,7 +2209,7 @@ ValueDefault=false
 
 [Indent Namespace Level]
 Category=2
-Description="<html>The number of spaces to indent a namespace block.</html>"
+Description="<html>The number of spaces to indent a namespace block.<br/>If set to zero, use the value indent_columns</html>"
 Enabled=false
 EditorType=numeric
 CallName="indent_namespace_level="
