+++ tests/config/Issue_1813.cfg
@@ -1,0 +1,11 @@
+indent_columns                  = 2
+indent_with_tabs                = 0
+indent_namespace                = true
+indent_namespace_single_indent  = true
+
+#The number of spaces to indent a namespace block.
+#indent_namespace_level         =
+
+#Whether to indent only the first namespace, and not any nested namespaces.
+#Requires indent_namespace=true.
+indent_namespace_single_indent  = true
