+++ tests/expected/cpp/30908-Issue_1813-2.cpp
@@ -1,0 +1,11 @@
+namespace n1 {
+namespace n2 {
+
+  void func() {
+    another_func([]() {
+      return 42;
+    });
+  }
+
+}
+}
