+++ tests/input/cpp/Issue_1813-3.cpp
@@ -1,0 +1,13 @@
+namespace n1 {
+                 namespace n2 {
+      namespace n3 {
+
+        void func() {
+      another_func([]() {
+                                     return 42;
+          });
+    }
+
+    }
+    }
+    }
